fix(fs): harden Windows rooted replacement checks
