+++ internal/fsdurable/errors.go
@@ -3,7 +3,7 @@ package fsdurable
 import "errors"
 
 var (
-	// ErrReplacementCommitted reports that a file replacement completed but a
-	// subsequent durability or verification operation failed.
+	// ErrReplacementCommitted reports that the rooted rename succeeded but a
+	// subsequent identity verification or durability operation failed.
 	ErrReplacementCommitted = errors.New("file replacement committed with uncertain durability")
 )
